sessions 0.3.0 developed

+++ blog/views/articles/v-one.php
@@ -6,10 +6,12 @@
     </div>
     <div><?= $article['content'] ?></div>
     <hr>
-    <div class="gap10">
-      <a class="btn btn-primary" href="<?= BASE_URL ?>edit/<?= $id ?>">Edit</a>
-      <a class="btn btn-danger" href="<?= BASE_URL ?>delete/<?= $id ?>">Remove</a>
-    </div>
+    <? if (isset($_SESSION['admin'])) : ?>
+      <div class="gap10">
+        <a class="btn btn-primary" href="<?= BASE_URL ?>edit/<?= $id ?>">Edit</a>
+        <a class="btn btn-danger" href="<?= BASE_URL ?>delete/<?= $id ?>">Remove</a>
+      </div>
+    <? endif; ?>
   </div>
 <? else : header($_SERVER['SERVER_PROTOCOL'] . ' 404 Not Found');
   $pageContent = render('v_404'); ?>
